requirements.txt
matplotlib
numpy
scipy
PySimpleGUI
PyYAML
scikit-learn